Changes according to #22
Moved the compression classes to a compression package.

Moved the compression tests to a compression test package.

Added some more lines to the compression / decompression test cases.

Remove one obsolete test for decompression strings.

Use the String Bash Escaping service to properly escape filenames for
compression and decompression.
